requirements.txt
# Core dependencies
numpy>=1.22.0
pandas>=1.4.0
scikit-learn>=1.0.2
scipy>=1.8.0
matplotlib>=3.5.1
pillow>=9.0.1

# Deep learning
torch>=1.10.0
torchvision>=0.11.1
transformers>=4.16.2

# LightGBM
lightgbm>=3.3.2

# CLIP
ftfy>=6.1.1
regex>=2022.1.18

# Text processing
emoji>=1.7.0
spacy>=3.2.0

# API
fastapi>=0.75.0
uvicorn>=0.17.0
python-multipart>=0.0.5
pydantic>=1.9.0

# Logging and monitoring
python-json-logger>=2.0.2

# Testing
pytest>=7.0.0
httpx>=0.22.0 